Send email

+++ app/Http/Controllers/JobController.php
@@ -2,39 +2,22 @@
 
 namespace App\Http\Controllers;
 
+use App\Mail\JobPosted;
 use App\Models\Job;
+use App\Models\User;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Mail;
 
 class JobController extends Controller
 {
     public function index()
     {
-        //orderBy("id", "asc") nach ID sortieren aufsteigend
-        $jobs = Job::with("employer")->latest()->simplePaginate(4);
-        return view('jobs/index', [
-            'jobs' => $jobs]);
-    }
+        $jobs = Job::with('employer')->latest()->simplePaginate(3);
 
-    public function show(Job $job)
-    {
-        return view("jobs.show", ["job" => $job]);
-    }
-
-    public function store()
-    {
-        request()->validate([
-            "title" => ["required", "min:3"],
-            "salary" => ["required"],
-
+        return view('jobs.index', [
+            'jobs' => $jobs
         ]);
-
-        Job::create([
-            "title" => request("title"),
-            "salary" => request("salary"),
-            "employer_id" => 1
-        ]);
-        return redirect("/jobs");
     }
 
     public function create()
@@ -42,28 +25,59 @@ class JobController extends Controller
         return view('jobs.create');
     }
 
+    public function show(Job $job)
+    {
+        return view('jobs.show', ['job' => $job]);
+    }
+
+    public function store()
+    {
+        request()->validate([
+            'title' => ['required', 'min:3'],
+            'salary' => ['required']
+        ]);
+
+        $job = Job::create([
+            'title' => request('title'),
+            'salary' => request('salary'),
+            'employer_id' => 1
+        ]);
+
+        Mail::to($job->employer->user)->send(
+            new JobPosted($job)
+        );
+
+        return redirect('/jobs');
+    }
+
     public function edit(Job $job)
     {
-        return view("jobs.edit", ["job" => $job]);
+        return view('jobs.edit', ['job' => $job]);
     }
 
     public function update(Job $job)
     {
+        Gate::authorize('edit-job', $job);
+
         request()->validate([
-            "title" => ["required", "min:3"],
-            "salary" => ["required"]
+            'title' => ['required', 'min:3'],
+            'salary' => ['required']
         ]);
 
         $job->update([
-            "title" => request("title"),
-            "salary" => request("salary"),
+            'title' => request('title'),
+            'salary' => request('salary'),
         ]);
-        return redirect("/jobs/" . $job->id);
+
+        return redirect('/jobs/' . $job->id);
     }
 
     public function destroy(Job $job)
     {
+        Gate::authorize('edit-job', $job);
+
         $job->delete();
-        return redirect("/jobs");
+
+        return redirect('/jobs');
     }
 }
